Refactor:  move all slice files of your Redux into directories
