+++ src/Pages/Missions.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import Header from '../Components/Header';
-import { fetchMissions, selectMissions, joinMission } from '../app/features/missionReducer';
+import { fetchMissions, selectMissions, joinMission } from '../redux/mission/missionReducer';
 import styles from './Missions.module.css';
 
 const Missions = () => {
